feat: split start sync logic
