+++ libs/jwst-rpc/src/lib.rs
@@ -1,13 +1,15 @@
 mod broadcast;
 mod channel;
+mod client;
 
 pub use channel::Channels;
+pub use client::start_client;
 
 use axum::extract::ws::{Message, WebSocket};
 use broadcast::subscribe;
 use channel::ChannelItem;
 use futures::{sink::SinkExt, stream::StreamExt};
-use jwst::{error, info, trace};
+use jwst::{debug, error, info, trace, warn};
 use jwst_storage::JwstStorage;
 use std::sync::Arc;
 use tokio::sync::mpsc::channel;
