+++ libs/jwst/src/lib.rs
@@ -11,7 +11,7 @@ pub use history::{
     parse_history, parse_history_client, BlockHistory, HistoryOperation, RawHistory,
 };
 pub use log::{debug, error, info, trace, warn};
-pub use types::{BlobMetadata, BlobStorage, DocStorage, DocSync, JwstError, JwstResult};
+pub use types::{BlobMetadata, BlobStorage, DocStorage, JwstError, JwstResult};
 pub use utils::sync_encode_update;
 pub use workspaces::{MapSubscription, Workspace, WorkspaceTransaction};
 #[cfg(feature = "workspace-search")]
